Added basic methods, need to implement. Been researching and looking at how to pull from multiple file paths, will be implementing them in the next commit. I thought I could 'gate' the functionality to work for both post- iOS8 and pre-iOS8. However this code is pretty ugly and the compiler is freaking out on me, it works, but god it sucks. Refactor: I am going to split the UIWebView and the WKWebView methods into two different classes of View Contoller for readability and usability sake. The pre-compiler directives are confusing and unnecessary.
After I figure out the best way to deal with this, I'll be a lot more productive. I learned the facts I need.

I am going to create a front-end service.
Unfortunately, iOS does not support node as it does JIT through Nitro. However Nitro is underdeveloped in terms of libraries in iOS, thus I am going to focus on a front-end service and explore this in more detail with less of a time crunch.